+++ agents-flex-store/agents-flex-store-pgvector/src/main/java/com/agentsflex/store/pgvector/PgvectorUtil.java
@@ -16,12 +16,14 @@
 package com.agentsflex.store.pgvector;
 
 import org.postgresql.util.PGobject;
+
 import java.sql.SQLException;
 
 public class PgvectorUtil {
     /**
      * 转化为vector.
      * 如果需要half vector或者sparse vector 对应实现即可
+     *
      * @param src 向量
      * @return
      * @throws SQLException
@@ -44,16 +46,16 @@ public class PgvectorUtil {
         return vector;
     }
 
-    public static double[] fromPgVector(String src) {
+    public static float[] fromPgVector(String src) {
         if (src.equals("[]")) {
-            return new double[0];
+            return new float[0];
         }
 
         String[] strs = src.substring(1, src.length() - 1).split(",");
-        double[] output = new double[strs.length];
+        float[] output = new float[strs.length];
         for (int i = 0; i < strs.length; i++) {
             try {
-                output[i] = Double.parseDouble(strs[i]);
+                output[i] = Float.parseFloat(strs[i]);
             } catch (Exception ignore) {
                 output[i] = 0;
             }
